backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-anymail[sendgrid]==10.3
drf-spectacular==0.27.2
pywebpush==2.0.0
psycopg[binary]==3.1.19
gunicorn==22.0.0
python-json-logger==2.0.7
django-ratelimit==4.1.0
whitenoise==6.6.0

pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
ruff==0.4.4
